first successful attempt at unpacking AZFP format
This is a commit of the first successful attempt at reading AZFP.  The data is not being plotted correctly but I'm pretty confident it is being extracted.  Right now it does not spit out or save the unpacked data, it only unpacks it and attempts to plot which it writes out as a PNG then exits.