Update test_ecs_integration.py (#1575)
* Update test_ecs_integration.py

Recent xarray versions (2025.11.0) now preserve attributes by default. Because of this, .identical() fails even when values and coordinates are identical, which broke several ECS tests. We now use .equals(), which ignores attributes and keeps the tests stable.

* Update requirements.txt

Update the code so that all tests and RTD builds pass, following the changes from @ctuguinay
Ref: Unpin jupyter-book in docs/requirements.txt #1570